+++ src/main/java/org/appjam/bongbaek/domain/member/service/MemberService.java
@@ -11,9 +11,9 @@ import org.appjam.bongbaek.global.common.CommonErrorCode;
 import org.appjam.bongbaek.global.exception.CustomException;
 import org.appjam.bongbaek.global.jwt.dto.TokenResponse;
 import org.appjam.bongbaek.global.jwt.dto.TokenResponse.Token;
-import org.appjam.bongbaek.global.jwt.util.JwtParser;
-import org.appjam.bongbaek.global.jwt.util.JwtProvider;
-import org.appjam.bongbaek.global.jwt.util.JwtValidator;
+import org.appjam.bongbaek.global.jwt.components.JwtParser;
+import org.appjam.bongbaek.global.jwt.components.JwtProvider;
+import org.appjam.bongbaek.global.jwt.components.JwtValidator;
 import org.appjam.bongbaek.global.oauth.kakao.KakaoLoginClient;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
