+++ src/main/java/org/appjam/bongbaek/global/jwt/filter/JwtAuthenticationFilter.java
@@ -7,8 +7,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.appjam.bongbaek.global.jwt.util.JwtParser;
-import org.appjam.bongbaek.global.jwt.util.JwtValidator;
+import org.appjam.bongbaek.global.jwt.components.JwtParser;
+import org.appjam.bongbaek.global.jwt.components.JwtValidator;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
