--- src/main/java/org/appjam/bongbaek/global/jwt/util/JwtValidator.java
@@ -1,41 +1,0 @@
-package org.appjam.bongbaek.global.jwt.util;
-
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.UnsupportedJwtException;
-import lombok.RequiredArgsConstructor;
-import org.appjam.bongbaek.global.common.CommonErrorCode;
-import org.appjam.bongbaek.global.exception.CustomException;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class JwtValidator {
-
-    private final JwtParser jwtParser;
-
-    /**
-     * 토큰 유효성 검증
-     * - 성공하면 true 반환
-     * - 실패하면 CustomException 바로 던짐
-     */
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser()
-                .verifyWith(jwtParser.getSigningKey())
-                .build()
-                .parseSignedClaims(token);
-
-            return true;
-        } catch (SecurityException e) {
-            throw new CustomException(CommonErrorCode.INVALID_ACCESS_TOKEN);
-        } catch (MalformedJwtException e) {
-            throw new CustomException(CommonErrorCode.INVALID_ACCESS_TOKEN);
-        } catch (ExpiredJwtException e) {
-            throw new CustomException(CommonErrorCode.INVALID_ACCESS_TOKEN);
-        } catch (UnsupportedJwtException e) {
-            throw new CustomException(CommonErrorCode.INVALID_ACCESS_TOKEN);
-        }
-    }
-}
